requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
bitarray==2.9.2
black==24.1.1
build==1.0.3
cbor2==5.6.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cytoolz==0.12.3
dataclassy==0.11.1
eip712==0.2.4
eth-account==0.10.0
eth-brownie==1.20.1
eth-event==1.2.5
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.0.0
execnet==2.0.2
frozenlist==1.4.1
hexbytes==0.3.1
hypothesis==6.27.3
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
lru-dict==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
parsimonious==0.9.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
py==1.11.0
py-solc-ast==1.2.10
py-solc-x==1.1.1
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pygments-lexer-solidity==0.7.0
pyproject_hooks==1.0.0
pytest==6.2.5
pytest-forked==1.6.0
pytest-xdist==1.34.0
python-dotenv==0.16.0
pyunormalize==15.1.0
pywin32==306
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.17.1
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
tinyec==0.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
vvm==0.1.0
vyper==0.3.10
wcwidth==0.2.13
web3==6.15.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
